+++ components/Dex.js
@@ -1,30 +1,33 @@
-import React from "react";
-import { useHistory } from "react-router-dom";
-import { withRouter } from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import Router from 'next/router';
 
 const Dex = (props) => {
+  const [selected, isSelected] = useState(false);
 
   function capitalizeFirstLetter(string) {
     return string.charAt(0).toUpperCase() + string.slice(1);
   }
-  
-  const history = useHistory();
 
-  function redirect(props) {
-    history.push({
-      pathname: '/exchanges',
-      state: {
-        dex: `${props.name}`,
+  const selectDex = () => {
+    isSelected(true);
+  }
+
+  useEffect(() => {
+    if (selected) {
+      const {pathname} = Router
+      if(pathname == '/' ){
+          // Router.push('/exchange')
+          Router.push({
+          pathname: '/exchange',
+          query: { id: props.name},
+        })
       }
-    });
-  }  
+    }
+ }, [selected]);
 
   return (
     <>
-      <tr className="row" 
-        onClick={() => redirect(props)}
-
-      >
+      <tr className="row" onClick={selectDex}>
         <td className="dex-id">{props.id}</td>
         <td className="dex-name">{capitalizeFirstLetter(props.name)}</td>
         <td className="dex-img"><img srcSet={props.image} alt="dex-img" /></td>
@@ -35,4 +38,4 @@ const Dex = (props) => {
   );
 };
 
-export default withRouter(Dex);
+export default Dex;
