+++ components/selectAccount.js
@@ -1,10 +1,9 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import Grid from '@material-ui/core/Grid';
 import SearchIcon from '@material-ui/icons/Search';
 import { alpha, makeStyles } from '@material-ui/core/styles';
 import InputBase from '@material-ui/core/InputBase';
 import NewAccount from './newAccount';
-import { useEffect } from 'react';
 
 const useStyles = makeStyles((theme) => ({
     root: {
