+++ next.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    reactStrictMode: true,
+    images: {
+        domains: ['duneanalytics.com'],
+    },
+}
