+++ package.json
@@ -1,59 +1,37 @@
 {
-  "name": "harmony-scan",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@apollo/client": "^3.3.21",
-    "@cubejs-client/core": "^0.27.48",
-    "@cubejs-client/react": "^0.27.48",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@harmony-js/core": "^0.1.57",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.58",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.1",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
-    "chart.js": "^2.9.3",
-    "clsx": "^1.1.1",
-    "graphql": "^15.5.1",
-    "lodash": "^4.17.21",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.6.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "rsuite": "^4.10.2",
-    "tslib": "^2.3.0",
-    "web-vitals": "^1.1.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+    "name": "harmony-scan",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "@fortawesome/fontawesome-svg-core": "^1.2.35",
+        "@fortawesome/free-brands-svg-icons": "^5.15.3",
+        "@fortawesome/free-solid-svg-icons": "^5.15.3",
+        "@fortawesome/react-fontawesome": "^0.1.14",
+        "@harmony-js/core": "^0.1.57",
+        "@material-ui/core": "^4.12.2",
+        "@material-ui/icons": "^4.11.2",
+        "@material-ui/lab": "^4.0.0-alpha.58",
+        "apollo-server-micro": "^3.0.2",
+        "axios": "^0.21.1",
+        "chart.js": "^2.9.3",
+        "clsx": "^1.1.1",
+        "graphql": "^15.5.1",
+        "micro": "^9.3.4",
+        "next": "11.0.1",
+        "prop-types": "^15.7.2",
+        "react": "17.0.2",
+        "react-dom": "17.0.2",
+        "swr": "^0.5.6"
+    },
+    "devDependencies": {
+        "eslint": "7.31.0",
+        "eslint-config-next": "11.0.1"
+    },
+    "license": "MIT"
 }
