+++ pages/api/graphql.js
@@ -1,0 +1,30 @@
+// import { ApolloServer, gql } from 'apollo-server-micro'
+
+// const typeDefs = gql`
+//   type Query {
+//     users: [User!]!
+//   }
+//   type User {
+//     name: String
+//   }
+// `
+
+// const resolvers = {
+//   Query: {
+//     users(parent, args, context) {
+//       return [{ name: 'Nextjs' }]
+//     },
+//   },
+// }
+
+// const apolloServer = new ApolloServer({ typeDefs, resolvers })
+
+// export const config = {
+//   api: {
+//     bodyParser: false,
+//   },
+// }
+
+// await apolloServer.start();
+
+// export default apolloServer.createHandler({ path: '/api/graphql' })
