+++ pages/api/hello.js
@@ -1,0 +1,5 @@
+// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+
+// export default function handler(req, res) {
+//   res.status(200).json({ name: 'John Doe' })
+// }
