--- src/components/HistoryChart/HistoryChart.css
@@ -1,21 +1,0 @@
-.margin {
-    margin: 8px !important;
-}
-
-.extendedIcon {
-    margin-right: 8px;
-}
-
-.charting {
-    background: #f8f8ff;
-    border-color: rgba(174, 305, 194, 0.4);
-    margin-left: 200px;
-    margin-right: 200px;
-    padding-left: 80px;
-    padding-right: 80px;
-    padding-top: 32px;
-    padding-bottom: 32px;
-    color: #00AEE9;
-    border: 5px solid;
-    border-radius: 25px;
-}
