--- src/components/PairChart/PairChart.css
@@ -1,9 +1,0 @@
-.charting {
-    background: #f8f8ff;
-    border-color: #00AEE9;
-    margin-left: 200px;
-    margin-right: 200px;
-    color: #00AEE9;
-    border: 5px solid;
-    border-radius: 25px;
-}
